+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 
 ### 1.1.0
     - Added functions for file storage
+
+### 1.1.3
+    - Now supporting PHP-Rocker 1.1.0
+    - Added method removeFile(name, callback)